+++ t/apicast-policy-rate-limit.t
@@ -83,7 +83,7 @@ Return 200 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("42_connections_test1")
+      redis:del('42_connections_test1')
     }
   }
 
@@ -159,7 +159,7 @@ Return 200 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("42_connections_test2")
+      redis:del('42_connections_test2')
     }
   }
 
@@ -277,7 +277,7 @@ Return 200 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("connections_test4")
+      redis:del('connections_test4')
     }
   }
 
@@ -394,7 +394,8 @@ Return 200 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("leaky_bucket_test6_1", "connections_test6_2", "fixed_window_test6_3")
+      local redis_key = redis:keys('*_fixed_window_test6_3')[1]
+      redis:del('leaky_bucket_test6_1', 'connections_test6_2', redis_key)
     }
   }
 
@@ -469,7 +470,7 @@ Return 429 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("connections_test7")
+      redis:del('connections_test7')
     }
   }
 
@@ -529,7 +530,7 @@ Return 503 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("leaky_bucket_test8")
+      redis:del('leaky_bucket_test8')
     }
   }
 
@@ -589,7 +590,8 @@ Return 429 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("fixed_window_test9")
+      local redis_key = redis:keys('*_fixed_window_test9')[1]
+      redis:del(redis_key)
     }
   }
 
@@ -667,7 +669,7 @@ Return 200 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("connections_test10")
+      redis:del('connections_test10')
     }
   }
 
@@ -728,7 +730,7 @@ Return 200 code.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("leaky_bucket_test11")
+      redis:del('leaky_bucket_test11')
     }
   }
 
@@ -1065,7 +1067,9 @@ so only the third call returns 429.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("fixed_window_test17_1", "fixed_window_test17_2")
+      local redis_key1 = redis:keys('*_fixed_window_test17_1')[1]
+      local redis_key2 = redis:keys('*_fixed_window_test17_2')[1]
+      redis:del(redis_key1, redis_key2)
     }
   }
 
@@ -1153,7 +1157,9 @@ so only the third call returns 429.
       local redis = require('resty.redis'):new()
       redis:connect(redis_host, redis_port)
       redis:select(1)
-      redis:del("fixed_window_localhost/test18_1", "fixed_window_localhost/test18_2")
+      local redis_key1 = redis:keys('*_fixed_window_localhost/test18_1')[1]
+      local redis_key2 = redis:keys('*_fixed_window_localhost/test18_2')[1]
+      redis:del(redis_key1, redis_key2)
     }
   }